feature : add solution 55 leetcode

+++ cses/missing-number/Solution.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Solution {
    public static void main(String[] args) {
       Scanner sc = new Scanner(System.in);
-      int n = sc.nextInt();
+      long n = sc.nextLong();
 
       long sum = (n * (n + 1)) / 2;
       long counter = 0;
