+++ cses/number-spiral/main.cpp
@@ -1,16 +1,18 @@
 #include <iostream>
-#define ll long 
+#define ll long long int
 
 using namespace std;
 
-int main() {
-    int t;
-    cin >> t;
-    while(t--) {
-      long a, b,mx,mn;
-      cin >> a >> b;
-      mx = max(a,b);
-      mn = min(a,b);
-      cout << (mx*mx) - (mn-1) << endl;
+int main()
+{
+  ll t;
+  cin >> t;
+  while (t--)
+  {
+    ll a, b, mx, mn;
+    cin >> a >> b;
+    mx = max(a, b);
+    mn = min(a, b);
+    cout << (mx * mx) - (mn - 1) << endl;
   }
 }
